+++ database/migrations/2022_08_06_210439_create_pet_registers_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(Pet::class);
             $table->text('register');
-            $table->unsignedSmallInteger('type');
+            $table->tinyInteger('type');
             $table->timestamps();
         });
     }
